% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hmdget.R
\name{HMDargs}
\alias{HMDargs}
\title{Parse Arguments of \code{HMDget} Function}
\usage{
HMDargs(.x)
}
\arguments{
\item{.x}{\code{\link{HMDget}} arguments}
}
\value{
Parsed arguments
}
\description{
Parse Arguments of \code{HMDget} Function
}
\details{
This functions does error checking and translates the function
  arguments of \code{\link{HMDget}} into the needed format.
}

