% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hmdget.R
\name{HMDhttp}
\alias{HMDhttp}
\title{Download HMD data from the web}
\usage{
HMDhttp(.x, .measure, .username, .password)
}
\arguments{
\item{.x}{Country + Timeframe design matrix [data frame]}

\item{.measure}{"Dx" (death counts) or "Nx" (exposures
in person years) [string, length == 1]}

\item{.username}{HMD username [string, length == 1]}

\item{.password}{HMD password [string, length == 1]}
}
\value{
HMD web data [data frame]
}
\description{
Download HMD data from the web
}

