% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hmdget.R
\docType{data}
\name{hmdcbook}
\alias{hmdcbook}
\title{HMD Country Codebook}
\format{A data frame with 46 rows and 2 variables: \describe{
  \item{Code}{Country code as used in Human Mortality Database}
  \item{Label}{Full country name in English} }}
\source{
\url{http://www.mortality.org/cgi-bin/hmd/hmd_download.php}
}
\usage{
hmdcbook
}
\description{
A dataset containing country codes and corresponing labels as used in the HMD
database and webpage. Use it to specify the countries you want to download
from the HMD database. \code{\link{HMDget}} only allows for country codes
specified in this table in its \code{.country} argument.
}
\keyword{datasets}

