% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hmdget.R
\name{HMDget}
\alias{HMDget}
\title{Download HMD Data from the Web}
\usage{
HMDget(.country, .timeframe = "p+c", .measure, .username, .password)
}
\arguments{
\item{.country}{HMD country code as found in \code{\link{hmdcbook}}, multiple
countries allowed}

\item{.timeframe}{"c" (cohort), "p" (period)  or "p+c" (period and cohort)}

\item{.measure}{"Dx" (death counts), "Nx" (exposures in person years) or "mx"
(mortality rates)}

\item{.username}{HMD username}

\item{.password}{HMD password}
}
\value{
Death counts, exposures or mortality rates by country, timeframe,
  year, age and sex in long format with numeric age categories.
}
\description{
Download HMD Data from the Web
}
\examples{
\dontrun{
HMDget(.country   = "SWE",
       .timeframe = "p",
       .measure   = "mx",
       .username  = "***",
       .password  = "***")
}
}

