% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hmdget.R
\name{HMDget}
\alias{HMDget}
\title{Download HMD data from the web}
\usage{
HMDget(.country, .timeframe = "p+c", .measure, .username, .password)
}
\arguments{
\item{.country}{3 letter country code [string, length >
0]}

\item{.timeframe}{"c" (cohort), "p" (period)  or "p+c"
(period and cohort) [string, length == 1]}

\item{.measure}{"Dx" (death counts) or "Nx" (exposures
in person years) [string, length == 1]}

\item{.username}{HMD username [string, length == 1]}

\item{.password}{HMD password [string, length == 1]}
}
\value{
Death counts or exposures by country, timeframe,
  year, age and sex in long format with numeric age
  categories [data frame].
}
\description{
Download HMD data from the web
}

