% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hmdget.R
\name{HMDhttp}
\alias{HMDhttp}
\title{Connect to HMD Server and Download Data}
\usage{
HMDhttp(.x, .measure, .username, .password)
}
\arguments{
\item{.x}{Country + Timeframe design matrix}

\item{.measure}{"Dx" (death counts), "Nx" (exposures in person years) or "mx"
(mortality rates)}

\item{.username}{HMD username}

\item{.password}{HMD password}
}
\value{
HMD web data
}
\description{
Connect to HMD Server and Download Data
}

