% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hmdget.R
\name{HMDtidy}
\alias{HMDtidy}
\title{Tidy HMD web data}
\usage{
HMDtidy(.x, .measure)
}
\arguments{
\item{.x}{HMD web data [data frame]}

\item{.measure}{"Dx" (death counts) or "Nx" (exposures
in person years) [string, length == 1]}
}
\value{
Tidied HMD data [data frame]
}
\description{
Tidy HMD web data
}
\details{
This function takes HMD data as downloaded from
  the web, reshapes it into long format (\code{Female},
  \code{Male}, \code{Total} columns to \code{Sex} +
  \code{Value} column) and converts the age variable to
  numeric (and in the process changing the age category
  \code{110+} to 110).
}

